[FIX/#125] 검색, 홈뷰 repository 통합
