+++ app/src/main/java/co/orange/ddanzi/di/module/DataSourceModule.kt
@@ -7,7 +7,6 @@ import co.orange.data.dataSource.HomeDataSource
 import co.orange.data.dataSource.IamportDataSource
 import co.orange.data.dataSource.InterestDataSource
 import co.orange.data.dataSource.ProfileDataSource
-import co.orange.data.dataSource.SearchDataSource
 import co.orange.data.dataSource.SellDataSource
 import co.orange.data.dataSource.SettingDataSource
 import co.orange.data.dataSourceImpl.AuthDataSourceImpl
@@ -17,7 +16,6 @@ import co.orange.data.dataSourceImpl.HomeDataSourceImpl
 import co.orange.data.dataSourceImpl.IamportDataSourceImpl
 import co.orange.data.dataSourceImpl.InterestDataSourceImpl
 import co.orange.data.dataSourceImpl.ProfileDataSourceImpl
-import co.orange.data.dataSourceImpl.SearchDataSourceImpl
 import co.orange.data.dataSourceImpl.SellDateSourceImpl
 import co.orange.data.dataSourceImpl.SettingDataSourceImpl
 import dagger.Module
@@ -40,10 +38,6 @@ object DataSourceModule {
     @Provides
     @Singleton
     fun provideDeviceDataSource(deviceDataSourceImpl: DeviceDataSourceImpl): DeviceDataSource = deviceDataSourceImpl
-
-    @Provides
-    @Singleton
-    fun provideSearchDataSource(searchDataSourceImpl: SearchDataSourceImpl): SearchDataSource = searchDataSourceImpl
 
     @Provides
     @Singleton
