+++ app/src/main/java/co/orange/ddanzi/di/module/RepositoryModule.kt
@@ -7,7 +7,6 @@ import co.orange.data.repositoryImpl.HomeRepositoryImpl
 import co.orange.data.repositoryImpl.IamportRepositoryImpl
 import co.orange.data.repositoryImpl.InterestRepositoryImpl
 import co.orange.data.repositoryImpl.ProfileRepositoryImpl
-import co.orange.data.repositoryImpl.SearchRepositoryImpl
 import co.orange.data.repositoryImpl.SellRepositoryImpl
 import co.orange.data.repositoryImpl.SettingRepositoryImpl
 import co.orange.data.repositoryImpl.UploadRepositoryImpl
@@ -18,7 +17,6 @@ import co.orange.domain.repository.HomeRepository
 import co.orange.domain.repository.IamportRepository
 import co.orange.domain.repository.InterestRepository
 import co.orange.domain.repository.ProfileRepository
-import co.orange.domain.repository.SearchRepository
 import co.orange.domain.repository.SellRepository
 import co.orange.domain.repository.SettingRepository
 import co.orange.domain.repository.UploadRepository
@@ -42,10 +40,6 @@ object RepositoryModule {
     @Provides
     @Singleton
     fun provideDeviceRepository(deviceRepositoryImpl: DeviceRepositoryImpl): DeviceRepository = deviceRepositoryImpl
-
-    @Provides
-    @Singleton
-    fun provideSearchRepository(searchRepositoryImpl: SearchRepositoryImpl): SearchRepository = searchRepositoryImpl
 
     @Provides
     @Singleton
