+++ data/src/main/java/co/orange/data/dataSource/HomeDataSource.kt
@@ -2,7 +2,10 @@ package co.orange.data.dataSource
 
 import co.orange.data.dto.BaseResponse
 import co.orange.data.dto.response.HomeDto
+import co.orange.data.dto.response.SearchResultDto
 
 interface HomeDataSource {
     suspend fun getHomeData(): BaseResponse<HomeDto>
+
+    suspend fun getSearchResult(keyword: String): BaseResponse<SearchResultDto>
 }
