--- data/src/main/java/co/orange/data/dataSource/SearchDataSource.kt
@@ -1,8 +1,0 @@
-package co.orange.data.dataSource
-
-import co.orange.data.dto.BaseResponse
-import co.orange.data.dto.response.SearchResultDto
-
-interface SearchDataSource {
-    suspend fun getSearchResult(keyword: String): BaseResponse<SearchResultDto>
-}
