+++ data/src/main/java/co/orange/data/dataSourceImpl/HomeDataSourceImpl.kt
@@ -3,6 +3,7 @@ package co.orange.data.dataSourceImpl
 import co.orange.data.dataSource.HomeDataSource
 import co.orange.data.dto.BaseResponse
 import co.orange.data.dto.response.HomeDto
+import co.orange.data.dto.response.SearchResultDto
 import co.orange.data.service.HomeService
 import javax.inject.Inject
 
@@ -12,4 +13,6 @@ data class HomeDataSourceImpl
         private val homeService: HomeService,
     ) : HomeDataSource {
         override suspend fun getHomeData(): BaseResponse<HomeDto> = homeService.getHomeData()
+
+        override suspend fun getSearchResult(keyword: String): BaseResponse<SearchResultDto> = homeService.getSearchResult(keyword)
     }
