+++ data/src/main/java/co/orange/data/repositoryImpl/HomeRepositoryImpl.kt
@@ -2,6 +2,7 @@ package co.orange.data.repositoryImpl
 
 import co.orange.data.dataSource.HomeDataSource
 import co.orange.domain.entity.response.HomeModel
+import co.orange.domain.entity.response.SearchResultModel
 import co.orange.domain.repository.HomeRepository
 import javax.inject.Inject
 
@@ -11,4 +12,9 @@ class HomeRepositoryImpl
         private val homeDataSource: HomeDataSource,
     ) : HomeRepository {
         override suspend fun getHomeData(): Result<HomeModel> = runCatching { homeDataSource.getHomeData().data.toModel() }
+
+        override suspend fun getSearchResult(keyword: String): Result<SearchResultModel> =
+            runCatching {
+                homeDataSource.getSearchResult(keyword).data.toModel()
+            }
     }
