--- domain/src/main/kotlin/co/orange/domain/repository/SearchRepository.kt
@@ -1,7 +1,0 @@
-package co.orange.domain.repository
-
-import co.orange.domain.entity.response.SearchResultModel
-
-interface SearchRepository {
-    suspend fun getSearchResult(keyword: String): Result<SearchResultModel>
-}
